multiconfigs: Change CONFIG_DTFILE to PATH + basename
The CONFIG_DTFILE_DIR is defined in the machine.conf file, while all of
the generated DTS files live in that directory.  Instead of hard coding
the full path, only specify the base name + the path variable with is
already excluded from the vardeps.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>

Acked-for-series: Sandeep Gundlupet Raju <sandeep.gundlupet-raju@amd.com>