*_flow.py: ParseCpuDict: Add check for enabled multilibs
When there is no self.MultiConfUser or self.MultiConfMap, then the user
is not expecting to generate multiconfig and related files.  While this
configuration is standard in xsct, it can also happen in SDT if the user
disabled absolutely everything.  (We may also want this configuration
to be available in the future for "generic" OS builds.)

The check inserted is the same check that was already present in the
multiconfig.py function that did similar processing.

This fixes the error identified in xsct_flow.py:
  [ERROR] Unable to find selected multiconfig ()

Signed-off-by: Mark Hatle <mark.hatle@amd.com>

Acked by : Ashwini Lomate <ashwini.lomate@amd.com>