yocto_machine.py: Switch to new external-hdf and sdt-artifacts
The URI of the hardware configuration is now expected to be stored in the
machine file.  This obsoletes HDF_EXAMPLES and similar 'single point
of reference' hardware artifact files.

The advantage of this method is the URI used to create the machine
is the one the machine will use.  Much less chance of a mismatch.

With the new code, references to the bitstream, pdi file, etc have been
changed to to reference sdt-artifacts and the WORKDIR where the
components would be.  This eliminates a further reference to something
the Yocto Project might not be aware of for checksumming.

These values are AFTER the require of the inherited config.  This ensures
that these settings always take priority.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>