gen-machineconf: Change sub parser logic to better detect flow
validate_hwfile is modified to only process the appropriate file if the user
specified the flow.  Otherwise, it will fall back to a dynamic detection
behavior.  This helps fix a problem where the user passes in a specific
flow type ends up with confusing error messages.

Also change from != to ==, this makes it clear we're only checking for
things the user has enabled instead of assuming things based on what the
user has NOT enabled.  The logic will help if we ever introduce additional
workflows in the future.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>

Acked-for-series: Sandeep Gundlupet Raju <sandeep.gundlupet-raju@amd.com>