sdt_flow.py: Change PSU_INIT_PATH processing
We only defined PSU_INIT_PATH if the user sets the parameter.  Ottherwise
the recipe will default the PSU_INIT_PATH to be the same as the
SYSTEM_DTFILE_PATH.  (Implemented in the layer.)

Additionally the user is warned that setting the psu-init-path will result
in a non-portable machine .conf being generated.  This may be useful for
development, but the SDT artifacts should be corrected with the right
ps init files for a portable build.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>

Acked-for-series: Sandeep Gundlupet Raju <sandeep.gundlupet-raju@amd.com>