yocto_machine.py: Update machine conf generation
The machine conf now specifies if it is an XSA or SDT workflow.  This enables
configuration where both meta-xilinx-tools and meta-xilinx-standalone-sdt are
included at the same time.  Additionally, XSA builds also add in the XSCT
version to avoid mismatches between ESW and XSCT.

Some of the generated items were reordered a bit.  This ensures proper
variable definiton behavior within the parsing of the machine conf files.

Additionally strip the trailing slash off of any generated path names to
avoid various issues.  For instance the system will warn if the 'S' value ends
with a '/' as it can cause packaging bugs.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>

Acked-for-series: Sandeep Gundlupet Raju <sandeep.gundlupet-raju@amd.com>