Fix errors when bitbake can not be run due to an error
In some cases, it might happen that bitbake, while present, can not be run
due to an internal error.  (Often sanity check of TMPDIR will trigger a
failure.)  In this case, drop back to the non-bitbake behavior.  If the files
are local this should allow things to complete.

Additionally translate any direct path names into file with absolute/full
path names.

Finally switch the bb.fetch2.FetchError exception to common_utils.FetchError.
We need to be able to distinguish between a general 'bitbake error' and a
specific FetchError.  If a FetchError occurs, we tell the user the fetch failed,
otherwise we try to fall back to the copy behavior.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>
Acked-by: Raju Kumar Pothuraju <rajukumar.pothuraju@amd.com>