common_utils.py: Skip actions when bitbake is not available
In some cases bitbake is not available (i.e. TMPDIR on an NFS partition).  We
previously attempted to detect this and fail gracefully, however there were
additional places where the components could attempt to do bitbake actions
that would fail.

Additionally the prepare step, when trying to get recipe values, was found
to have the wrong parsing argument, this has been corrected as well.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>
Acked-by: Sandeep Gundlupet Raju <sandeep.gunldupet-raju@amd.com>