gen-machineconf: Fix issue where system-top.dts is used
When system-top.dts is passed in, the system would download the ONE file and
then try to process it.  All of the associate incldue and dtsi files were
missed.

Try to detect this special case.  Is the file local to the disk?  Is the
name system-top.dts, and does the directory it's in contain an include
directory?  If so, we change the full name to the path insted.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>
Reviewed-by: Sandeep Gundlupet Raju <sandeep.gunldupet-raju@amd.com>
Acked-by: Raju Kumar Pothuraju <rajukumar.pothuraju@amd.com>