Revise the way and when bitbake is started up
Instead of multiple bitbake starting points, we always start it (and activate
the parser) in a try block.  If bitbake won't start or we make sure it
shutdowns and is disabled for any further steps.

Since BBPATH processing requires bitbake, we move all processing to where we
are checking for the BBPATH.

Finally we add an argument to startBitbake that allows us to instantiate the
class, with everything disabled.  This will permit us (in the future) to
globally disable bitbake, if necessary.

getVar has also been improved to catch exceptions and return back 'None', this
will shield us in some cases where parsing can not be finished due to an errors
such as TMPDIR is on NFS.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>
Reviewed-by: Sandeep Gundlupet Raju <sandeep.gundlupet-raju@amd.com>
Acked-by: Varalaxmi Bingi <Varalaxmi.bingi@amd.com>